TVMultiSARMA

Documentation for TVMultiSARMA.

TVMultiSARMA.ConvertWideMat2Vec

TVMultiSARMA.FindActiveLagsMultiSAR

TVMultiSARMA.GibbsLocalMultiSAR

TVMultiSARMA.MultiSARtoReg

TVMultiSARMA.NormalApproxUniformStationary

TVMultiSARMA.PlotCompanionEigenMultiSAR

TVMultiSARMA.PlotMultiSAREvolution

TVMultiSARMA.PostSpecDensMultiSAR

TVMultiSARMA.SARMAasReg

TVMultiSARMA.SetupARReg

TVMultiSARMA.plotEvolStabilityRegion

TVMultiSARMA.ConvertWideMat2Vec — Method

ConvertWideMat2Vec(ϕARpost, p, season)

Converts a T × sum(p) × nIter matrix to a length(season) vector with T × p[l] × nIter elements for the l:th season with lag order p[l]

TVMultiSARMA.FindActiveLagsMultiSAR — Method

FindActiveLagsMultiSAR(p, s)

Find the active lags in a general multi-seasonal SAR(p,s) process with seasons in the vector s. The number of lags for each seasonal polynomial is given in the vector p.

TVMultiSARMA.GibbsLocalMultiSAR — Function

GibbsLocalMultiSAR(x, modelSettings, priorSettings, algoSettings)

Gibbs sampling from the posterior of the locally stable time-varying multi-seasonal AR(p) model with dynamic shrinkage process prior. Optionally, with stochastic volatility.

TVMultiSARMA.MultiSARtoReg — Method

MultiSARtoReg(θ::Vector, p::Vector, s::Vector, activeLags; pacf_map = "monahan", 
    negative_signs = true)

Takes a vector θ of length sum(p) with unrestricted AR/SAR coefficients and returns the non-zero coefficients (determined by activeLags) in the product polynomial Πₗ(1 - ϕₗ₁B^(sₗ) - ϕₗ₂B^(2*sₗ) - ....) = (1 - ϕ̃₁B - ϕ̃₂B² - ....) where the parameters in each AR polynomial is restricted to stability region.

TVMultiSARMA.NormalApproxUniformStationary — Method

μ₀, Σ₀ = NormalApproxUniformStationary(p)

Returns mean vector (μ₀) and covariance matrix (Σ₀) in the θ ~ Normal approximation to the uniform prior for the unrestricted θₖ in the multi-seasonal AR with p regular lags and seasonal lags.

For example, p = [2,2,1] is SAR model with two regular lags, two lags at the first seasonal period (e.g. daily cycle) and a single lag on the second seasonal period (e.g. yearly cycle). The approximation minimizes the Hellinger distance to the t and skew-t priors.

Examples

julia> μ₀, Σ₀ = NormalApproxUniformStationary([2, 1])
([0.0, -0.529625, 0.0], [1.0857014809 0.0 0.0; 0.0 0.7363356099999999 0.0; 0.0 0.0 1.0857014809])

TVMultiSARMA.PlotCompanionEigenMultiSAR — Method

PlotCompanionEigen(ϕevol)

Computes eigenvalues of the companion matrix for an evolving AR process

TVMultiSARMA.PlotMultiSAREvolution — Method

PlotMultiSAREvolution(x, ϕevol, θevol, p, season)

Plot the time series and the evolution of the multi-seasonal AR parameters ϕ, θ over time.

TVMultiSARMA.PostSpecDensMultiSAR — Method

PostSpecDensMultiSAR(ϕARpost, σₑpost, p, season; ωgrid = 0.01:0.01:π, thinFactor = 100)

Compute posterior draws of the spectral density for a time-varying multi-season SAR model from posterior draws of the AR coefficients (ϕARpost a T×sum(p)×nIter matrix) at every thinFactor iteration.

TVMultiSARMA.SARMAasReg — Method

ϕ̂, sₑ = SARMAasReg(x, pFit, season, imposeZeros = true)
Fits a SARMA model as a regression and returns the OLS estimates of the AR coefficients and the standard deviation of the residuals.

TVMultiSARMA.SetupARReg — Method

y, Z, T = SetupARReg(x, p)

Sets up the lagged matrix Z and the response vector y for an AR(p) process.

The AR(p) process xt = ϕ₁*x{t-1} + ... + ϕp*x{t-p} + ν_t is transformed into the regression model:

y = Z*ϕ + ν

where y = [x{p+1}, x{p+2}, ..., xT], the t:th row of Z is Z[t,:] = [x{t-1}, x{t-2}, ..., x{t-p}] and ϕ = [ϕ₁, ϕ₂, ..., ϕ_p].

Note: we loose the first p observations in the transformation and the return T is adjusted accordingly to T-p.

Examples

julia> x = simARMA([0.7,-0.2], [0.0], 0.0, 1.0, 100)
julia> y, Z, T = SetupARReg(x, p)
julia> inv(Z'*Z)*Z'y

TVMultiSARMA.plotEvolStabilityRegion — Function

plotEvolStabilityRegion(ϕevol, θevol, iter, plotEvolLine = true)

Plot the parameter evoluation in θ-space and ϕ-space (triangle in AR(2) case) up to iteration iter. plotEvolLine = true plots a line connecting the last 10 iterations.